componentización controles
