+++ src/components/controles/Panelitems/panelItem.css
@@ -1,0 +1,53 @@
+.panel-item {
+  display: flex;
+  justify-content: space-between; /* Alinear texto e interruptor */
+  align-items: center; /* Centrar verticalmente */
+  margin-bottom: 10px;
+  font-size: 14px; /* Tamaño de fuente */
+}
+.switch {
+  position: relative;
+  display: inline-block;
+  width: 34px;
+  height: 20px;
+}
+.switch input {
+  opacity: 0;
+  width: 0;
+  height: 0;
+}
+
+.slider {
+  position: absolute;
+  cursor: pointer;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #ccc;
+  transition: 0.4s;
+  border-radius: 20px;
+}
+.slider:before {
+  position: absolute;
+  content: "";
+  height: 14px;
+  width: 14px;
+  left: 3px;
+  bottom: 3px;
+  background-color: white;
+  transition: 0.4s;
+  border-radius: 50%;
+}
+
+input:checked + .slider {
+  background-color: #91cb92; /* Verde cuando está activado */
+}
+input:checked + .slider:before {
+  transform: translateX(14px);
+}
+
+span{
+  font-family: 'Times New Roman', Times, serif;
+  font-size: medium;
+}
