+++ src/components/controles/ThemeButtons/themeButton.css
@@ -1,0 +1,22 @@
+.theme-buttons{
+  display: flex;
+  display: inline;
+  justify-content: content-column;
+  flex-direction: column;
+  margin-top: 5px;
+  margin-bottom: 10px;
+}
+
+.theme-select {
+  background-color: rgba(55, 65, 81, 0.8); /* Sin fondo */
+  border: 2px solid #bdbdbd;
+  border-radius: 8px;
+  padding: 8px 12px;
+  font-size: 16px;
+  color: #ffffff;
+}
+
+.theme-select option {
+  background-color: rgba(55, 65, 81, 0.8); /* Sin fondo */
+  color: #f4ecec;
+}
