+++ src/components/controles/controles.css
@@ -1,19 +1,3 @@
-.settings-container {
-  position: relative;
-}
-
-.settings-icon {
-  position: fixed;
-  bottom: 25px;
-  right: 0;
-  background-color: rgba(55, 65, 81, 0.8); /* Sin fondo */
-  color: white;
-  padding: 8px;
-  border-radius: 8px 0 0 8px;
-  cursor: pointer;
- 
-}
-
 .settings-panel {
   position: absolute;
   bottom: 105px;
@@ -38,84 +22,4 @@
   list-style: none;
   padding: 0;
   margin: 0;
-}
-
-.panel-item {
-  display: flex;
-  justify-content: space-between; /* Alinear texto e interruptor */
-  align-items: center; /* Centrar verticalmente */
-  margin-bottom: 10px;
-  font-size: 14px; /* Tamaño de fuente */
-}
-
-.switch {
-  position: relative;
-  display: inline-block;
-  width: 34px;
-  height: 20px;
-}
-
-.switch input {
-  opacity: 0;
-  width: 0;
-  height: 0;
-}
-
-.slider {
-  position: absolute;
-  cursor: pointer;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: #ccc;
-  transition: 0.4s;
-  border-radius: 20px;
-}
-
-.slider:before {
-  position: absolute;
-  content: "";
-  height: 14px;
-  width: 14px;
-  left: 3px;
-  bottom: 3px;
-  background-color: white;
-  transition: 0.4s;
-  border-radius: 50%;
-}
-
-input:checked + .slider {
-  background-color: #91cb92; /* Verde cuando está activado */
-}
-
-input:checked + .slider:before {
-  transform: translateX(14px);
-}
-
-span{
-  font-family: 'Times New Roman', Times, serif;
-  font-size: medium;
-}
-.theme-buttons{
-  display: flex;
-  display: inline;
-  justify-content: content-column;
-  flex-direction: column;
-  margin-top: 5px;
-  margin-bottom: 10px;
-}
-
-.theme-select {
-  background-color: rgba(55, 65, 81, 0.8); /* Sin fondo */
-  border: 2px solid #bdbdbd;
-  border-radius: 8px;
-  padding: 8px 12px;
-  font-size: 16px;
-  color: #ffffff;
-}
-
-.theme-select option {
-  background-color: rgba(55, 65, 81, 0.8); /* Sin fondo */
-  color: #f4ecec;
 }
